Time zone display in long instead of short

+++ src/course-home/outline-tab/SequenceLink.jsx
@@ -52,7 +52,7 @@ const SequenceLink = ({
             day="numeric"
             month="short"
             year="numeric"
-            timeZoneName="short"
+            timeZoneName="long"
             value={due}
             {...timezoneFormatArgs}
           />
